refactor(webapp): split rewards distribution state
